+++ Sources/Frontend/Scan.swift
@@ -46,6 +46,9 @@ public final class Scan: Injectable {
 
         try Analyzer.perform(graph: graph)
 
-        return ScanResult(declarations: graph.dereferencedDeclarations, graph: graph)
+        let reducer = RedundantDeclarationReducer(declarations: graph.dereferencedDeclarations)
+        let reducedDeclarations = reducer.reduce()
+
+        return ScanResult(declarations: reducedDeclarations, graph: graph)
     }
 }